backend/requirements.txt
flask
flask-cors
PyGithub
# Add any other dependencies your backend needs

